Do commit check after adding federations
Was modeled after cluster adding where commit the check was being done
before sending the changes to the dbd. But since the dbd isn't making
any additional tables for federations -- like it does for clusters -- it
can send to the dbd first and make sure they worked and then ask to
commit the changes.